Add files via upload
Written a bunch of methods, Mainly, I worked on the make() method, as well as several methods that will be called by it. This method lets the used add new persons and events in the program.

Some problems are apparent. New persons added during the running of the programs are either not instantiated, or at least not stored as the program closes. There is also the issue that currently all interests a person can have must be only a single word with no spaces in them. Need to fix both of these.