Add files via upload
Fixed a bug on instantiating a new event object.
- Problem: I was adding integer values to the string array eventStrings.
- Fix: Now these values go to the integer array eventStartTimeUnits.

Made the invitation status value in the invitation HashMap in the Event class an enum, and made the InvitationStatus enum class to match.

Wrote all the remaining core and supporting methods to make editing invitations for new and existing events possible. After testing it looks like all /make functionality works as it should.

Started on the /edit functionality in the InvitationBook class, which is the class containing the main method.