Add files via upload
Been working on some of the last functionality to be added: about(), as well as functionality that lets users manage interests in a more sweeping way than just adding/changing/removing them one person at a time. Both of these are unfinished, but I believe I have landed on the way I want to do it.